01-DocumentCleanup/requirements.txt
tqdm==4.46.1
torch==1.1.0
torchvision==0.3.0
opencv-python==4.4.0.46
opencv_contrib_python==4.4.0.46
imageio==2.8.0
scipy==1.4.1
pandas==1.0.3
numpy==1.18.1
matplotlib==3.1.3
Shapely==1.7.0
Pillow==8.0.1
scikit_learn==0.22.1
